Fixed up some styles and changed date display issues.
