--- client/src/config/const.js
@@ -1,2 +1,0 @@
-export const baseUrl =
-  process.env.NODE_ENV === "development" ? "http://localhost:3001/api" : "";
